Introduce cp token attestor (#68)
This is the first pass at a spire server token-based node attestor. Currently, it only supports loading tokens by config file. The tokens are one-time-use only, and have a TTL applied. I realize that's fairly limiting... we will beef this up though with time, first we need to get attestation working end-to-end.

I had to update the common protobuf definition too, somehow we changed a type during that fix. I also had to touch some of the grpc glue code... not sure if it's right or wrong, but it's working :P

* Delete old placeholder plugins

* Introduce CP join_token node attestor
First pass at simple config-driven join token node attestor for control plane

* Fix up CP Node Attestor join token plugin following dir restructuring

* Clean up server attestor grpc package

* Fix km_memory plugin config mixup

* Use a mutex in server join token plugin
https://golang.org/ref/mem