Simple change to add go:generate comments to run mockgen
Add //go:generate comments to the source files of the  *_mock.go files. This way, they can be easily generated just running "go generate", or with a click from the IDE. Also, is easier to figure out that the files generate code that needs to be updated if the source file changes.
- https://blog.golang.org/generate