Switch to registered join tokens With CLI token registration exposed and datastore updated to persist them, switch the node api over to use the registered tokens instead of the configured ones.
Also expose join token flag on agent, which supersedes any configured node attestor plugin

Note that this commit changes the format of the base spiffe id from `spiffe/node-id/*` to `spiffe/node/TYPE/*`. In practice, node IDs are not human friendly, so including the type in the path makes it clear what kind of ID it is, and gives each one its own namespace.