Update base spiffe id format We have declared base spiffe ID format to be `spiffe://example.org/spiffe/node-id/*` where `*` is a unique node id. In practice, however, this format is cumbersome. It is very difficult to tell the type of an ID from just looking at it, particularly so when the ID is a GUID. Since the IDs are naturally typed in the attestedData message, it makes sense to encode the type in the base SPIFFE ID.
Furthermore, the term `node` can be ambiguous, particularly in situations where the agent is run in a pod or container. To add insult to injury, the `spiffe` basename has led to confusion about whether this is part of the SPIFFE spec, or whether it is SPIRE-specific.

This commit updates the format to `spiffe://example.org/spire/agent/TYPE/*` where `TYPE` is the type of attested data, i.e. join_token, aws, etc. Pending SIG-Components review