Introducing statically-linked plugins This commit teaches the common catalog to handle "builtin" plugins. Builtin plugins are, for all intents and purposes, identical to plugins run out-of-process: they conform to the same interfaces, are configured the same way, and are consumed from the catalog in the same way. For the consumer, the experience is identical. For users, they will no longer need to worry about managing multiple processes. As more plugin options are introduced, the user can choose which to enable by simply configuring them as enabled, similar to the way they would do with an out-of-process plugin.
Agent and server specific catalogs will understand which plugins should be linked in, generating references to them and passing them into the common catalog. This commit updates the server catalog to statically link all existing checked-in plugins it relies upon. A subsequent PR will introduce similar changes in the agent's catalog.

I will note that with statically-linked plugins, a number of previously-held conventions no longer make sense... for instance, we could probably do without the `enabled` option, and perhaps the format of the configuration file could benefit from a refactor. Additionally, the catalog has undergone a number of changes in recent months (including this one), and could itself benefit from a refactor to clean it up a bit. Finally, the testing methodologies used for the various plugins are a bit out-of-place, and I have had to make some minor updates to plugin instantion and testing code in order to get everything working smoothly. I have stopped short of actually attempting any of these refactors in order to keep this change well scoped.

The top-level plugin directory doesn't make much sense for builtins, so I have moved them to a new subdirectory in the server package. I have also done some light renaming in order for things to be a little more idiomatic.

I expect at least a couple refactors to follow this. The goal of this change was to introduce statically-linked plugins without boiling the oceans, and a lot of the spots needing refactor (like CA plugin tests) were in need of refactors prior to this change anyways.

Now, let's get that AWS plugin set checked in :).