Add certificate pruner to server CA manager
This code is meant to "clean up" old CA certificates which have since expired. It provides a 24h grace period before doing so, and will refuse to prune the full set. This code will be heavily amended in the coming days in order to support safer rotation methods, but wanted to get this work checked in sooner rather than later.