Check in the AWS IID node attestor This commit checks in the node attestor work done by @juturnas in github.com/spiffe/aws-iid-attestor
It has essentially been copy/pasted in the interest of time, applying only the necessary changes requred for a built-in plugin (hence the absence of tests).

The plugin name for this builtin is `aws_iid`, as opposed to the out-of-process plugin name of `aws_iid_attestor`... Besides being more idiomatic, the new name should prevent collisions with the previous plugin should it be loaded on top of an updated codebase. Though the code is identical, I have not performed an end-to-end test due to lack of AWS creds - it would be great if @walmav or @drrt could give this a quick whirl to ensure everything is good.

Once #369 is merged, I will either update this PR or send a new PR to add documentation for this addition.