Rotate SPIRE CA when expiry is approaching This change add support for CA rotation when an expiration date is approaching. When the server's signing certificate is >50% old, a new certificate is prepared and injected into the bundle. When the signing certificate is approx. >80% old, the new certificate is activated and installed for signing.
Note that we currently do not have a system by which we can ensure it is safe to rotate. That is, we don't have a good way to know if the new certificate has been fully propagated. This is a problem we may or may not have to solve somewhere down the line... for now, we will count on workload API push and push/short-poll from agent. Small steps :)