Check in client for workload API This commit introduces a Workload API client, implementing backoff as needed. It exposes an update channel to its consumers which is decoupled from the updates delivered by the Workload API - this is intended to prevent poorly-behaving client code from blocking Workload API updates being sent by a node agent. Node agents should implement mitigations for slow clients too, but that is out of scope for this change.
This is being checked into the SPIRE project instead of go-spiffe for a few reasons. The first is that it is difficult to consume the protobuf code from SPIRE since the SPIRE project is vendored, leading to type mismatches in non-vendored consumers like the go-spiffe library. Second is that it makes sense for client code to be in the same repository as the API definition to which it is conforming. We can follow this up by pulling in the rest of the go-spiffe functionality - I have created an `api` top-level directory for such libraries, following established convention.