Add telemetry to SPIRE agent This commit introduces a common package for telemetry data. There was some back-and-forth on whether or not to use the exported interfaces from go-metrics or to implement our own. We originally thought that we should not duplicate the code, however in instrumenting the code I found that there was a convenience function missing from the interface that I'd prefer be included. Additionally, implementing our own interface for telemetry will allow us to add custom stuff like transparent prefix and label addition... not to mention an easier transition to other libraries if we choose.
As a result, this commit 1) introduces a common telemetry package, complete with SPIRE-specific interfaces and wrap code, and 2) takes the first steps towards instrumenting the SPIRE agent.

I have not yet hooked up CLI or configurables for statsd destination, but I have implmented them in the agent code. Sending the agent SIGUSR1 will trigger a statistics dump. Once we figure out how we would like to surface this data, we can hook up the CLI and config for the desired method with little effort. A similar PR will follow to implement server instrumentation.

I'm not completely happy with the current state, and feel that some additional helpers may make sense here, but for now this will do. There are already a few surprises in the data :).