Update gRPC Dialer to take a func for transport credentials After a disconnection, the gRPC library will (by default) attempt to reconnect. In doing this however, it can obscure some of the underlying error conditions which cause the disconnect in the first place. Martin did some work to introduce a custom dialer which exposes a logger, so that we get to know the reason for dial failures.
As an add-on to that work, it's desirable to pick up the latest client credentials on every retry. A disconnect and subsequent retry on an existing dialer using old credentials will result in a permanent error rather than a recovery. In order to ensure that the dialer always has up-to-date credentials, this commit switches the GRPCDialer to accept a func which return active credentials, rather than accepting static credentials at the time of instantiation. This should allow for seamless client credential rotation _and_ the builtin gRPC retries.

In doing this work, I have found that the cache manager attempts to handle these kinds of failures already... I am not sure if that logic works when the connection is severed or not, but I left that stuff in place, and instead pass a static-ish function. I suspect that with future refactors, we will be able to better utilize this update.

Signed-off-by: Evan Gilman <evan@scytale.io>