Ensure gRPC error messages are logged at the error level
Pull #437 switched the grpcutil dialer to use a stdlib logger interface instead of the logrus field logger. This is desirable because we want to use the dialer in many places, not all of which have access to a full-blown logrus instance (e.g. the cli utilities).

This introduces a new problem though: logrus `Print` methods are hardcoded to log at info level, and these are the methods which are compat with the standard logger. Anything coming from the gRPC dialer should be an error and not info level.

This commit works around this issue by creating standard logger instances backed by a logrus io writer which is hardcoded to write at the error level instad of the info level.

Signed-off-by: Evan Gilman <evan@scytale.io>