increase resilience finding pod containing pid
If attestation is requested while the pod/containers are still
initializing, the containerID identified for the pid (via cgroups) may
not be present in the container status.

This change introduces a retry loop. Retries are only attempted if the
desired container is not found and there exists one or more status
entries that do not have the containerID present.

The max attempts and polling interval on the retry loop can be
controlled via the plugin configuration.

Signed-off-by: Andrew Harding <azdagron@gmail.com>