protoc-gen-spireplugin and dummy plugin
protoc-gen-spireplugin is a protoc plugin for generating spire plugin
interfaces from grpc service descriptions. It provides client and server
interfaces that can be uniformly consumed without the need to
distinguish between in-process (builtin) and out-of-process plugin
implementations. This is particularly important for plugins utilizing
streaming APIs, since the GRPC generated interfaces are asymmetric.

Adapters are provided that:
1) Wrap built-in plugin implementations on the client side so they can
be consumed in-process.
2) Wrap out-of-proc plugin implementations on the server side so they
conform to the generated GRPC interface.

Signed-off-by: Andrew Harding <azdagron@gmail.com>