x509 proof-of-possession node attestation
This adds a generic x509 node attestor. The server verifies that the
presented certificate chains back to a configured set of trusted roots.
It then performs proof-of-possession of the private key by issuing a
signature based challenge of a random nonce. The agent calculates the
signature and sends the response back to the server, at which point if
the signature verifies correctly, attestation is successful.

Signed-off-by: Andrew Harding <azdagron@gmail.com>