fix CA activation bug
The CA rotation check happens on an interval (default 1m). If the new CA
should be both prepared and rotated on the same check (due to a small
enough ttl) then there is a bug that would prepare on the first check
and wait to activate until the next check. The CA would expire between
the two checks.

This PR fixes the bug by allowing for both preparation and activation
during the same rotation check.

Signed-off-by: Andrew Harding <azdagron@gmail.com>