[481] Let agent dialer do DNS resolution In the beginning, spire-agent was configured with the IP address of spire-server. As a quick win to support DNS names, a small step was added to the command line package to perform a one-shot DNS resolution if a name was supplied, and configure the agent with the first result. Of course, this meant that if the name record was updated, we wouldn't pick up the change.
The proposed solution was to move the resolution step closer to the
dialer. @marcosy contributed work in this direction (thank you
@marcosy!). While reviewing, I realized a few things:
* Agent was being configured with net.Addr so it could support other
transport types in the future with ease. We were using UDS transport in
the tests, but were otherwise maintaining this ability without actually
needing or using it right now
* net.Addr doesn't want DNS names
* Golang dialer actually implements the desired behavior - it will
resolve records on our behalf, and fall through the addresses if there
is more than one

Rather than continue to support net.Addr on the agent, which was forcing
us to handle DNS resolution on our own, I opted to back it out and go
for a host:port string instead. This now gets passed all the way down to
the dialer. TCP is now the hardcoded transport. If we need to support
others in the future, we can introduce additional Dialer methods.

Tested using HA mode in docker-compose, fixes #481.

Signed-off-by: Evan Gilman <evan@scytale.io>