use SHA256 instead of XOR to combine x509pop nonces
XOR allows for a malicious agent to fake key possession by finding an
existing set of plaintext + signature generated by the key and choosing
a nonce that will produce the plaintext when XOR'd with the server
nonce.

Swapping to SHA256 prevents the attacker from choosing a nonce that
hashes to the plaintext.

Signed-off-by: Andrew Harding <azdagron@gmail.com>