fix journal_mode and foreign_keys support
`journal_mode` and `foreign_key` settings are per-connection in
sqlite3. The existing code sets the right PRAGMAs when the database is
first opened, but that only impacts the first connection in the
database/sql.(*DB) connection pool. When a new connection is created, it
will not have those settings, which could lead to unecessary errors in
the case of journal_mode (e.g. "database locked") or database
inconsistencies in the case of foreign_keys.

This change takes advantage of github.com/mattn/go-sqlite3 query value
support for enabling those features for each connection. The code now
parses the connection string, converts it into a file: URI if needed,
and adds the proper _journal_mode and _foreign_keys query values.

Signed-off-by: Andrew Harding <azdagron@gmail.com>