don't fail FetchBundle if bundle does not exist
The datastore was recently refactored to use Request/Response messages
uniformly. Before that refactor, FetchBundle returned a Bundle message
directly from the RPC and therefore had to fail if the bundle didn't
exist (since it couldn't return a nil response).

This commit changes the behavior of FetchBundle to match that of the
other Fetch methods. Namely, if the requested bundle is not found, the
call succeeds and returns an empty response object (i.e., the bundle
pointer will be nil).

All FetchBundle callers have been audited to make sure they handle an
nil bundle appropriately.

Signed-off-by: Andrew Harding <azdagron@gmail.com>