Fix keypair loading when B active but A unprepared
When loading keypair sets, the manager always loads A into the "current"
slot and B into the "next" slot and then performs a swap if B is
supposed to be active. In other words, B gets moved to "current".

When B has been activated but the next A has yet to be prepared, the
loading code has a bug and would not perform the swap.

This commit fixes the bug so that B is properly swapped into "current"
under this condition.

Signed-off-by: Andrew Harding <azdagron@gmail.com>