update cached registration entry on sync
The agent sync code ignores registration entries that are already in the
cache. This commit changes the behavior to update the cache if the
registration entry has changed (notifying subscribers).

This fix is preparatory to adding registration entry update support to
SPIRE server, which is currently unimplemented in the CLI and a stub in
the registration API.

Signed-off-by: Andrew Harding <azdagron@gmail.com>