Store CA intermediates in bundle
SPIRE was recently refactored to properly convey signing intermediates
down through the Node and Workload APIs. At that time the bundle was
changed to only include the signing root. Backwards compatability with
old agents was done at the Node API layer; intermediate certificates
from minted SVIDs were appended to the bundle before the result was
returned to the agent.

This approach was insufficient and broken since 1) the intermediates
would not be appended to bundles returned to agents talking with other
servers (in an HA scenario), and 2) the agent uses the same fetch call
to learn about new registration entries without asking for anything to
be signed, which caused the server to return a bundle without
intermediates.

This change removes the shim in the Node API and instead stores the
intermediates in the bundle so they are consistently returned from the
Node API and available in HA depoloyments. SPIRE 0.8.0 can revert the
change to no longer store the intermediates.

Signed-off-by: Andrew Harding <azdagron@gmail.com>