SecretDiscoveryService server implementation
This commit adds an implementation of the Envoy Secret Discovery Service
to the SPIRE agent. The SDS server is off by default and must be enabled
with the `enable_sds` config value.

Specifically, the SDS server is hosted on the same UDS socket as the
Workload API. It performs workload attestation to determine which
secrets to make available to the caller. Resources currently map as
follows:
- SVIDs map to TLSCertificate resources with a resource name matching
the SPIFFE ID of the SVID.
- Bundles map to ValidationContext resources with a resource name
matching the trust domain ID of the bundle.

At some point, we may consider symbolic names like
"default_tls_certificate" or "default_validation_context" that map to
certain resources, but that is out of scope for the initial
implementation.

Signed-off-by: Andrew Harding <azdagron@gmail.com>