Fix upstream CA handling
SPIRE server mishandles the upstream CA signing response by
misconstruing the UpstreamTrustBundle field as the certificate chain
from the certificate in the Cert field back to some trusted root. Only
the last certificate in the UpstreamTrustBundle is added to the trust
bundle, and the rest are considered intermediates. This prevents SPIRE
server from properly joining an upstream PKI.

This commit changes SPIRE server in the following ways:
1) UpstreamCA interface is changed to make it clear what belongs in the
response.  Namely, the CertChain field contains the newly signed CA
certificate and any intermediates necessary to chain back to a root that
exists in the UpstreamTrustBundle field. UpstreamTrustBundle is a set of
x509 roots for the upstream PKI. Compatability code remains in SPIRE
server to handle responses from old plugins.
2) Updates the CA manager to properly insert the upstream trust bundle
into the SPIRE server. The server CA is also added to the trust bundle
to maintain previous back-compat guarantees necessary until SPIRE 0.8.
If upstream_bundle is false, or if the upstream CA is not used, then the
server CA certificate continues to be the the only certificate added to
the trust bundle.
3) Adds some additional assertions to the upstream CA disk tests.

Signed-off-by: Andrew Harding <azdagron@gmail.com>