Introduce --node helper flag in spire-server cli Registration entries can apply to either nodes or workloads, but the behavior is slightly different for each case. The distinction between one and the other is made implicitly, and is difficult to understand without learning about SPIRE authz internals. This leads to a confusing experience, often to the point that users don't realize that registration entries can apply to nodes too.
While this problem deserves a more robust solution, this commit puts in
place a quick win that will hopefully serve to disambiguate the uses
between now and the time that we refine the model.

It adds a `--node` flag to the `spire-server entry create` command,
which indicates that the entry being created should be assigned to
nodes rather than workloads. When set, a Parent ID is no longer required
from the operator. Instead, the utility will set the Parent ID equal to
the server's SPIFFE ID under the covers.

Signed-off-by: Evan Gilman <evan@scytale.io>