build and publish docker images
This commit adds publishing of docker images to the gcr.io/spiffe-io
registry. Only builds from `push` events which are either tagged or from
the master branch are built and published. The images are published
under the commit hash. In addition, tagged builds
are published under the tag and master branch builds are published under
`unstable`.

Signed-off-by: Andrew Harding <azdagron@gmail.com>