revamped node handler tests
The node handler tests were very brittle, subtle, and hard to add new
test cases too. The overuse of mocks contributed to code that was very
hard to understand and update.

This commit rewrites the node handler tests to improve readability. Test
coverage has also been increased.

Also in this commit:
- cleaned up error messages
- added some additional validation of incoming requests (panic
mitigation)
- removed ability to sign a JWT for the agent SPIFFE ID (which didn't
work anyway due to assertions in the JWT signing support code).
- fixed fake server CA to imitate behavior of real server CA in
terms of what certificates belong to the returned chain. There is room
for future improvements and code reuse in this area.

Signed-off-by: Andrew Harding <azdagron@gmail.com>