fix sporadic node test failures
When testing bad authorization on streaming calls, the test code would
send a request on the stream first. Depending on timing, sending the
request would fail with EOF, since gRPC had already closed the stream
because of the expected auth failure. This would cause sporadic test
failures since we were checking the return code of Send().

Change up the auth tests to not send a request and only assert that
Recv() returns the expected error value.

Also added timeouts to all contexts passed into stream operations to
fail tests in a timely manner if something is broken.

Signed-off-by: Andrew Harding <azdagron@gmail.com>