K8S Bundle Notifier
This plugin publishes the latest trust bundle to a K8S ConfigMap when
SPIRE server is started and after bundle updates.

An IdentityProvider host service is also provided that the K8S Bundle
Notifier uses to fetch the bundle contents to perform conflict
resolution updating the ConfigMap using a read-modify-write cycle.

Signed-off-by: Andrew Harding <azdagron@gmail.com>