Fixed agent hang on shutdown closing UDS gRPC server
The UDS gRPC server credentials pull the peer PID off of the UDS.
Obtaining the FD for the connection was via the File() method.
Unfortunately, File() switches the underlying FD into blocking mode,
which prevents gRPC from closing the conns appropriately.

This changes up the credentials to instead obtain the FD (and do the
control call) via the syscall.RawConn. Doing so allows gRPC to close
down the connections appropriately, fixing the hang on
(*grpc.Server).Close()

Also removed the unnecessary call to close listeners, since the gRPC
server does that for us.

Signed-off-by: Andrew Harding <azdagron@gmail.com>