fix flaky tests around JWT issued at claims
The JWT signing code hard codes the use of time.Time() for crafting the
IAT claim. This cause spurious test failures in tests attempting to
assert the claim value is properly crafted.

This commit allows the JWT signing code to take an optional clock that
is used to craft the IAT claim. The change has some spidery consequences
that lead to more testing layers replacing the use of a "now" hook with
a clock.

Signed-off-by: Andrew Harding <azdagron@gmail.com>