Fix racy AWS IID test.
The recent catalog refactor introduced changes to tests wherein plugins
are loaded via the catalog builtin facilities during testing so that
loggers and hostservices and so on can be wired up the same way it
happens in production. This introduces subtle changes to behavior when
testing failure conditions to streaming RPCs that fail before the
request is sent. This fixes the AWS IID nodeattestor test to not fail
when the stream is torn down before it can send the request.

Signed-off-by: Andrew Harding <azdagron@gmail.com>