Fix the rotation test script I ran into two problems while trying to run the rotation test script. First, it did not work when an SVID with intermediate certificates was returned. To fix that, I added the `-untrusted` flag and simply pointed it at the same file that the SVID is in since it is also the file which contains the intermediates.
Second, we added keymanager disk to the default agent config, but the
script removes the datastore when it runs, which was causing the agent
authentication to fail as it wasn't known to the server. To fix that, I
updated the rotation script to remove some of the extra agent data that
is now persisted, which matches the e2e test script.

I also found that sqlite now keeps a couple extra files on disk that
will cause trouble if not cleaned up after a previous run. The e2e tests
weren't bothered by it because they run in travis only once, but it will
cause a problem on subsequent runs on, for instance, a dev machine.
Added those files to both e2e and rotation script for cleanup.

We should probably hook this into master merge at some point.

Signed-off-by: Evan Gilman <evan@scytale.io>