No longer require agents to determine the agent ID
This updates the attestation flow to no longer require agent-side
plugins to calculate and provide the agent ID. If an agent ID is
provided, previous behavior is maintained across the agent and server.

When a server receives a CSR with an agent ID, it still determines
"attested before" status that it passes into the server-side plugin. It
verifies that the agent ID returned from the server-side plugin matches
that in the CSR.

If the CSR does not contain an agent ID, the server relies on
server-side plugins that do not support re-attestation to check if the
agent is already known via a new AgentStore host service.

All of the builtin plugins are updated to conform to the new
expectations. A base implementation is also provided for node attestors to
embed to ease the transition.

The agent node attestor unit tests are also revamped as part of
deprecating the functionality due to poor coverage and difficult setup
(increased coverage from ~65% to ~90%).

Signed-off-by: Andrew Harding <azdagron@gmail.com>