agent/client: Finish in-progress requests before releasing
This changes the behavior of the agent's client.Release to not
immediately close the grpc connection, but to mark it for release
so that it can be closed after all in-progress FetchUpdates requests
have completed.

This allows uncoordinated callers in manager/sync and
rotator/svid to use the client concurrently, without the actions
of one negatively impacting the other.

Before this change:

1. The manager/sync module starts a FetchUpdates request.
2. The rotator/svid concurrently Releases the client being
   used by the manager.
3. The request in manager/sync fails immediately, since the
   underlying connection has been forcibly closed.

After this change:

1. The manager/sync module starts a FetchUpdates request.
2. The rotator/svid concurrently Releases the client being
   used by the manager, which schedules the old connection
   to be released.
3. The request in manager/sync run to completion.
4. The connection used in (3) is closed.

Signed-off-by: R. Tyler Julian <tjulian@uber.com>