Use the pared-down host service interface
Codegen for plugins, services, and host services provides a pared-down
client interface with the gRPC call options argument removed so the call
options cannot be abused by callers. The plugin and service codegen
provides the pared-down interface as the client interface but the host
service code does not.

This change updates the host service codegen to properly use the
pared-down interface as the client interface.

It also updates the codegen for the catalog test package to
automatically fix up imports (only necessary for testing).

Signed-off-by: Andrew Harding <azdagron@gmail.com>