Cut 0.8.0 This PR prepares us to cut SPIRE version 0.8.0, targeting commit c8123525fba8943bb084e70317afe32986247424.
Included is an updated CHANGELOG, which pulls in some history that was
kept in the v0.7 branch but never copied over to master. This PR also
bumps the current version to 0.8.1.

Once approved and merged, we will cut an annotated tag with the contents
of the CHANGELOG for this release. Unlike previous releases, we will
delay the creation of a version branch until we begin to introduce
changes intended for 0.9.0.

The release date in the CHANGELOG is tentatively set to Monday, May
20th.

Signed-off-by: Evan Gilman <evan@scytale.io>