Fix the UpstreamCA bootstrap use case In some cases, SPIRE users rely on an UpstreamCA to provide a stable root, but do not set upstream_bundle to true because they want the trust domain to be isolated from the upstream PKI. The stable root allows agents to be bootstrapped without having to obtain an up-to-date SPIRE bundle.
This use case broke with a recent CA refactor because the SPIRE server
TLS certificate does not serve a chain (i.e. it only serves the leaf)
when upstream_bundle is false. While this is the correct behavior under
normal conditions, it prevents the above stated use case from succeeding
because on first boot, agents don't have the CA certificate that the
server is managing - they only have the root that signed it.

I am not certain that we want to continue supporting this in the long
run. We should probably think through that. The default value of
upstream_bundle should also be switched to `true` in the near future as
well, since 1) it is surprising behavior to configure an UpstreamCA but
then not actually join the upstream PKI, 2) simply omitting higher
order certificates from the bundle is not a reliable isolation
mechanism, and 3) few X.509 validators will successfully validate a leaf
when the chain terminates with a non-self-signed root in the default
configuration (see openssl partial_chain).

This commit:
* Fixes the above stated use case by always including the server CA cert
as the second element when answering TLS requests
* Adds a functional test to ensure that this use case doesn't break in
the future
* Does NOT support this use case for any UpstreamCA where the signer is
not the root

Signed-off-by: Evan Gilman <evan@scytale.io>