fix flaky upstreamca csr test
The test case tests that a CSR with an invalid signature is rejected by
the upstream CA. It does so by crafting a CSR and then xor'ing the last
byte with itself. However, when the last byte is zero, the xor'ing
produces a zero, leaving a valid CSR.

Replace the xor operation with an increment (that will wrap).

Signed-off-by: Andrew Harding <azdagron@gmail.com>