Lightly refactor config merging in CLI run utilities The CLI utilities for running SPIRE server and agent were written in a way that makes it difficult to properly merge config inputs coming from file-based configuration and flag-based configuration. Several config structs are used to do this, but most importantly, there is one structure used for CLI-based config input (which holds mostly basic types), and another structure used for the actual core agent/server which holds more complicated types (like a logger).
Previously, the CLI logic would start with the core structure, and
progressively merge file-based CLI structure and flag-based CLI
structure on top, respectively. This made it difficult to do a 1:1
comparison because, for instance, you would have to know if the logging
format or level in the basic structure matched the already-configured
logger in the core structure. As a result, the code took some shortcuts.

This commit refactors the CLI code to instead merge all of the CLI
structures first, then convert the finalized result to the core
structure. This change also resolves a bug that prevented the user from
specifying one logging-specific config in a file (e.g. log format) and a
different logging-specific config by flag (e.g. log level).

Signed-off-by: Evan Gilman <evan@scytale.io>