Agent re-attests if its cached SVID is expired
Prior to this change, if an agent boots with an expired SVID in its
data directory, startup will fail with an error, for example:
"remote error: tls: bad certificate".

This commit adds a validity check when the SVID is loaded. If the
SVID is expired, it is ignored, and a new keypair is generated.

Signed-off-by: R. Tyler Julian <tjulian@uber.com>