datastore/sql: Add config for sql connection pools
Currently, the number of open connections to a database is
unbounded, as is the number of idle connections and the amount
of time a connection may be reused. These characteristics can
have an adverse effect on database performance.

This commit adds configuration to the datastore/sql plugin
so that users can tune common connection configuration to their
liking.

max_open_conns:
    description: The maximum number of open db connections
    default: 0 (unlimited)

max_idle_conns:
    use: The maximumn number of idle connections in the pool
    default: 2

conn_max_lifetime:
    use: The maximum amount of time a connection may be reused
    default: 0 (unlimited)

Signed-off-by: R. Tyler Julian <tjulian@uber.com>