Fix CLI config merging A bug was introduced in #866 which caused some default values to not take effect. Configurable precedence was also impacted. This was caught by the k8s integration suite as it relied on some of these defaults.
The problem ended up being twofold: the mergo lib treats the source
values as weaker than the destination values, and when WithOverride is
used, it is possible for non-zero values to be overwritten with zero
values.

I found the former to be surprising, and the latter should not happen
according to docs and code comments, but I did not have time to chase
down the exact reason it was occurring. Both of these problems were
addressed by reversing the order in which we merge configs such that the
"strongest" (i.e. highest precedence) values come first. Reversing the
order negates the need for WithOverride.

The `run` CLI commands and related configuration are poorly tested. We
were lucky to have caught this regression in our integration tests. To
ensure we don't run into this problem again, I've included a large
number of test cases in this commit which assert configurable
precedence, parsing, etc.

Also included in this commit is a minor change to an error message to
make it more accurate, and documentation additions for the missing
-logFormat CLI flag.

Signed-off-by: Evan Gilman <evan@scytale.io>